Implemented Game loop + End state
-Game now has a rough verison of a game loop in update method.

-Score and Time is tracked in imGui window.

-End Game Pop up appears when game ends.

Enemies now destroy along with bullets that hit them.

+++ DX11Starter/Enemy.cpp
@@ -15,7 +15,8 @@ Enemy::~Enemy()
 }
 
 //Return how many points this enemy is worth
-byte Enemy::GetPoints()
+int Enemy::GetPoints()
 {
 	return points;
 }
+
